pyproject.toml
[project]
name = "mobilityagent"
version = "0.1.0"
description = "AI-powered transportation planning system using CrewAI"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.0.0",
    "pandas>=2.2.3",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "crewai>=0.1.0",
    "crewai-tools>=0.32.1",
    "langchain>=0.1.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "ipython>=8.12.0",
    "requests>=2.31.0",
    "osmnx>=1.1.2",
    "geopandas>=0.10.2",
    "pyproj>=3.2.1",
    "geopy>=2.2.0",
    "snowflake-connector-python>=3.12.4",
]

run_crew = "mobilityagent.main:run"
mobilityagent = "mobilityagent.main:run"

scripts = { mobilityagent = "mobilityagent.main:main" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mobilityagent"]  # Replace "mobilityagent" with your actual package name
